Fix generator for apps and portfolios in lower editions

+++ test/unit/conftest.py
@@ -38,7 +38,6 @@ TEST_ISSUE = "a1fddba4-9e70-46c6-ac95-e815104ead59"
 
 def create_test_object(a_class: type, key: str) -> any:
     """Creates a SonarQube test object of a given class"""
-    util.start_logging()
     try:
         o = a_class.get_object(endpoint=util.SQ, key=key)
     except exceptions.ObjectNotFound:
@@ -70,55 +69,67 @@ def get_test_project() -> Generator[projects.Project]:
 @pytest.fixture
 def get_test_app() -> Generator[applications.Application]:
     """setup of tests"""
-    o = create_test_object(applications.Application, key=util.TEMP_KEY)
+    o = None
+    if util.SQ.version() in ("developer", "enterprise", "datacenter"):
+        o = create_test_object(applications.Application, key=util.TEMP_KEY)
     yield o
-    o.key = util.TEMP_KEY
-    try:
-        o.delete()
-    except exceptions.ObjectNotFound:
-        pass
+    if util.SQ.version() in ("developer", "enterprise", "datacenter"):
+        o.key = util.TEMP_KEY
+        try:
+            o.delete()
+        except exceptions.ObjectNotFound:
+            pass
 
 
 @pytest.fixture
 def get_test_portfolio() -> Generator[portfolios.Portfolio]:
     """setup of tests"""
-    o = create_test_object(portfolios.Portfolio, key=util.TEMP_KEY)
+    o = None
+    if util.SQ.version() in ("enterprise", "datacenter"):
+        o = create_test_object(portfolios.Portfolio, key=util.TEMP_KEY)
     yield o
-    o.key = util.TEMP_KEY
-    try:
-        o.delete()
-    except exceptions.ObjectNotFound:
-        pass
+    if util.SQ.version() in ("enterprise", "datacenter"):
+        o.key = util.TEMP_KEY
+        try:
+            o.delete()
+        except exceptions.ObjectNotFound:
+            pass
 
 
 @pytest.fixture
 def get_test_portfolio_2() -> Generator[portfolios.Portfolio]:
     """setup of tests"""
-    o = create_test_object(portfolios.Portfolio, key=util.TEMP_KEY_2)
+    o = None
+    if util.SQ.version() in ("enterprise", "datacenter"):
+        o = create_test_object(portfolios.Portfolio, key=util.TEMP_KEY_2)
     yield o
-    o.key = util.TEMP_KEY_2
-    try:
-        o.delete()
-    except exceptions.ObjectNotFound:
-        pass
+    if util.SQ.version() in ("enterprise", "datacenter"):
+        o.key = util.TEMP_KEY_2
+        try:
+            o.delete()
+        except exceptions.ObjectNotFound:
+            pass
 
 
 @pytest.fixture
 def get_test_subportfolio() -> Generator[portfolios.Portfolio]:
     """setup of tests"""
-    parent = create_test_object(portfolios.Portfolio, key=util.TEMP_KEY)
-    subp = parent.add_standard_subportfolio(key=util.TEMP_KEY_3, name=util.TEMP_KEY_3)
+    subp = None
+    if util.SQ.version() in ("enterprise", "datacenter"):
+        parent = create_test_object(portfolios.Portfolio, key=util.TEMP_KEY)
+        subp = parent.add_standard_subportfolio(key=util.TEMP_KEY_3, name=util.TEMP_KEY_3)
     yield subp
-    subp.key = util.TEMP_KEY_3
-    try:
-        subp.delete()
-    except exceptions.ObjectNotFound:
-        pass
-    parent.key = util.TEMP_KEY
-    try:
-        parent.delete()
-    except exceptions.ObjectNotFound:
-        pass
+    if util.SQ.version() in ("enterprise", "datacenter"):
+        subp.key = util.TEMP_KEY_3
+        try:
+            subp.delete()
+        except exceptions.ObjectNotFound:
+            pass
+        parent.key = util.TEMP_KEY
+        try:
+            parent.delete()
+        except exceptions.ObjectNotFound:
+            pass
 
 
 @pytest.fixture
@@ -215,21 +226,6 @@ def get_sarif_file() -> Generator[str]:
     yield file
     rm(file)
 
-
-@pytest.fixture
-def get_test_application() -> Generator[applications.Application]:
-    """setup of tests"""
-    util.start_logging()
-    try:
-        o = applications.Application.get_object(endpoint=util.SQ, key=util.TEMP_KEY)
-    except exceptions.ObjectNotFound:
-        o = applications.Application.create(endpoint=util.SQ, key=util.TEMP_KEY, name=util.TEMP_KEY)
-    yield o
-    try:
-        o.key = util.TEMP_KEY
-        o.delete()
-    except exceptions.ObjectNotFound:
-        pass
 
 
 @pytest.fixture
